refactoring useRichTextEditor
