+++ apps/web/app/composables/useRichTextEditor/helpers/expansion.ts
@@ -1,0 +1,14 @@
+import type { UseRichTextEditorArgs } from '../types';
+
+export const setupRichTextEditorExpansion = (args: UseRichTextEditorArgs) => {
+  const expandedLocal = ref<boolean>(args.expanded?.value ?? false);
+
+  if (args.expanded && args.onUpdateExpanded) {
+    watch(args.expanded, (val) => (expandedLocal.value = !!val));
+    watch(expandedLocal, (val) => args.onUpdateExpanded?.(val));
+  }
+
+  return {
+    expandedLocal,
+  };
+};
